fix: reject duplicate JFIF and Adobe app markers in JPEG
JPEG carrier file duplicated JFIF and Adobe app markers and if the
corresponding marker was not present in the source file, libjpeg
library functions jpeg_set_defaults() and jpeg_set_colorspace() set
JFIF or Adobe app marker, which changed metadata.

The goal of the StegoDisk library is to keep the JPEG metadata
unchanged (respectively, to modify only those parts of the JPEG
file in which information is inserted), so this patch fix it.